requirements.txt
boto3
click
pyyaml
diagrams
